+++ src/components/FileUpload/FileUpload.vue
@@ -40,8 +40,13 @@
 	const fileList = ref<File[]>([])
 	const isDarkMode = useTheme().current.value.dark
 
+	function openFileDialog() {
+		fileInput.value?.click()
+	}
+
 	defineExpose({
 		fileInput,
+		openFileDialog,
 	})
 
 	watch(() => props.modelValue, (value) => {
@@ -147,7 +152,7 @@
 .sy-file-upload {
 	cursor: pointer;
 	position: relative;
-	border: 1px dashed tokens.$colors-border-accent;
+	border: 1px dashed tokens.$colors-border-accent-primary;
 	border-radius: tokens.$radius-rounded-lg;
 	transition: background 0.25s;